@@ -6,14 +6,14 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.16.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   cupertino_icons: ^1.0.2
-  provider: ^5.0.0
+  provider:
   flutter_staggered_grid_view: ^0.4.0
   http: ^0.13.3
   flutter_snake_navigationbar: ^0.5.0
@@ -22,15 +22,15 @@ dependencies:
   fluttertoast: ^8.0.8
   hexcolor: ^2.0.4
   url_launcher: ^6.0.9
-  wallpaper_manager: ^1.0.10
   flutter_cache_manager: ^3.1.2
   image_gallery_saver: ^1.6.9
-  image_picker_saver:
   dio: ^4.0.0
-  permission_handler: ^8.1.4+2
-  image_downloader: ^0.20.1
-  shared_preferences: ^0.5.2
+  permission_handler: ^10.0.0
+  image_downloader: ^0.31.0
+  shared_preferences: ^2.0.15
   page_transition: "^2.0.4"
+  lottie: ^1.4.2
+  flutter_staggered_animations: ^1.0.0
 
 dev_dependencies:
   flutter_test:
@@ -40,7 +40,10 @@ flutter:
   uses-material-design: true
   assets:
     - assets/
-
+  fonts:
+    - family: DancingScript
+      fonts:
+        - asset: assets/fonts/DancingScript-Regular.ttf
   # To add assets to your application, add an assets section, like this:
 
 
